+++ src/graphql/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Public entry point for `nexusjs/graphql`.
+ */
+export * from "./types.js";
+export { GraphQLService, loadGraphQLJs } from "./graphql.service.js";
+export { GraphQLModule } from "./graphql.module.js";
+export {
+	Resolver,
+	Query,
+	Mutation,
+	Subscription,
+	Arg,
+	isResolverClass,
+	getResolverTypeName,
+	getResolverFields,
+	getMethodArgs,
+} from "./decorators/index.js";